refactor: decompose outcome_calculator + split market_data CLI
Part 1: Break 245-line _calculate_single_outcome into 3 focused helpers
(_resolve_base_price, _fill_timeframe_prices, _fill_intraday_prices)
plus a slim orchestrator.

Part 2: Split 754-line CLI into domain submodules (cli_fetch.py,
cli_outcomes.py, cli_registry.py) with cli.py as thin entry point
using add_command(). All 13 commands preserved with identical names.
